% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.timePlots.R
\name{bar.timePlots}
\alias{bar.timePlots}
\title{bar.timePlots}
\usage{
bar.timePlots(
  imputed.data,
  fav.proteins,
  fraction,
  timeSeries = FALSE,
  lfq.data,
  contrasts,
  sampleTable,
  pvalCutOff = 0.05,
  sigmaCutOff = 0.05,
  lfcCutOff = 0
)
}
\arguments{
\item{imputed.data}{Imputed data obtained after QC.filter}

\item{fav.proteins}{Vector with favourite proteins (or their corresponding sites) for which corresponding bar or timeseries plots are required}

\item{fraction}{Can be either "Proteome" or "Enriched". Indicates the type of input data used}

\item{timeSeries}{Logical (Default is FALSE). If TRUE, line plots will be generated instead of barplots}

\item{lfq.data}{Output from editData function}

\item{contrasts}{Mentions the conditions to be compared. Ex. MUTANT_vs_WILDTYPE or (MUTANT-A_vs_WILDTYPE-A)_vs_(MUTANT-B_vs_WILDTYPE-B). This how the contrasts should be provided and the values should be the same as the one given in condition column of sampleTable. (MUTANT-A_vs_WILDTYPE-A)_vs_(MUTANT-B_vs_WILDTYPE-B): This type can be used for complex data when comparing the interaction between two conditions such genotype and time}

\item{sampleTable}{.xlsx file containing information about the samples. Three columns are mandatory (label, condition and replicate)}

\item{pvalCutOff}{P-value cut off for significant protein(s), site(s) and GO terms. Default is 0.05}

\item{sigmaCutOff}{PI-value cut off for significant protein(s) and site(s). Default is 0.05 (Refer Xiao et al, 2014 and Hostrup et al, 2022 for details on PI-value)}

\item{lfcCutOff}{Log-fold change cut off. Default is 0}
}
\value{
Generates bar plot (or timeseries) for the selected protein(s) or site(s)
}
\description{
Generates barplots or timeseries plots
}
