% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editData.R
\name{editData}
\alias{editData}
\title{editData}
\usage{
editData(data, Fraction, probability, org, quantification)
}
\arguments{
\item{data}{Input or modified proteome file from MaxQuant (.txt)}

\item{Fraction}{Can be either "Proteome" or "Enriched". Indicates the type of input data used}

\item{probability}{Numeric value between 0-1. Filters out modified peptides with probabilities less than the given value (Only used if Fraction = "Enriched")}

\item{org}{Database of the organism. Drosophila melanogaster = "dme", Mus muscuslus ' "mmu", Homo sapiens = "hsa", Saccharomyces cerevisae = "sce". Default is "dme"}

\item{quantification}{Default is LFQ. Can be either "LFQ" or "iBAQ"}
}
\value{
A data frame that will be cleaner than the original raw data to be used for differential expression analysis
}
\description{
Cleans the raw data from MaxQuant for downstream analysis
}
\examples{
lfqdata <- editData(data = proteome, Fraction = "Enriched", probability = 0.9)

}
