% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.Data.R
\docType{data}
\name{enrich.Data}
\alias{enrich.Data}
\title{Modified proteome data}
\format{
An object of class \code{"data.frame"}
}
\source{
\href{http://www.ebi.ac.uk/pride/archive/projects/PXD042471}
}
\usage{
data(enrich.Data)
}
\description{
This file is obtained from MaxQuant as Acetyl(K)Sites.txt. Column information is not explained as these can be
obtained in the summary.pdf file from MaxQuant output This corresponding acetylome data is obtained from a
recently published study comparing the whole cell extracts of acetylomes between wild-type and chameau RNAi flies
(Further details can be found in the link provided below)
}
\examples{

data(enrich.Data)
head(enrich.Data)

}
\references{
Venkatasubramani AV, et al., (2023), EMBO Rep Oct 9;24(10):e57023
(\href{https://www.embopress.org/doi/full/10.15252/embr.202357023}{EMBO Reports})
}
\keyword{enriched}
