% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_normalization.R
\name{enrich_normalization}
\alias{enrich_normalization}
\title{enrich_normalization}
\usage{
enrich_normalization(
  protein.data,
  enrich.data,
  enrich.batch = c(TRUE, FALSE),
  probability
)
}
\arguments{
\item{protein.data}{Input proteome file from MaxQuant (.txt)}

\item{enrich.data}{Modified proteome file from MaxQuant (.txt)}

\item{enrich.batch}{If the normalization should be done based on paired replicates or average of replicates}

\item{probability}{Numeric value between 0-1. Filters out modified peptides with probabilities less than the given value}
}
\value{
Protein normalized modified-proteome data
}
\description{
Normalizes modified proteomes to the corresponding input proteomes
}
