% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentAnalysis.R
\name{EnrichmentAnalysis}
\alias{EnrichmentAnalysis}
\title{EnrichmentAnalysis}
\usage{
EnrichmentAnalysis(
  enrich = c("gsea", "ora"),
  nonExclusive.data = NULL,
  rankBy = "stat",
  KEGG = FALSE,
  ont = "BP",
  padjustMethod = "fdr",
  exclusive.data = NULL,
  background = TRUE,
  minGS = 50,
  maxGS = 500,
  simplify = FALSE,
  simplify_cutoff = 0.7
)
}
\arguments{
\item{enrich}{Can be either "gsea" or "enrich"}

\item{nonExclusive.data}{Default is NULL. If not, the data should be given as list containing data frame(s)}

\item{rankBy}{Can be either stat or logFC. Genes are ranked accordingly for GSEA}

\item{KEGG}{Logical, default is FALSE}

\item{ont}{Can be "BP", "CC", "MF" or "ALL"}

\item{padjustMethod}{Method for adjusting p-value (Refer clusterProfiler for available options)}

\item{exclusive.data}{Default is NULL. If not, the data should be given as list containing data frame(s)}

\item{background}{Logical, default is TRUE. If FALSE, background proteins will not be considered for over representation analysis}

\item{minGS}{minimal size of annotated genes in geneset}

\item{maxGS}{maximal size of annotated genes in geneset}

\item{simplify}{Logical, default is FALSE. If TRUE, redundant GO terms will be simplified based on the simplify_cutoff}

\item{simplify_cutoff}{Cutoff for simplify method. Numerical value from 0-1. Default is 0.7}
}
\value{
Generates list of significant GO terms
}
\description{
Performs Gene Set Enrichment Analysis and Over Representation Analaysis using clusterProfiler package
}
