% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\docType{data}
\name{fasta}
\alias{fasta}
\title{File from UNIPROT containing protein sequences from Drosophila melanogaster. Exact fasta file used also in MaxQuant search
Contains sequence information of all proteins from Drosophila melanogaster in fasta format. Obtained from UNIPROT}
\format{
An object of class \code{"character"}
}
\source{
Obtained from UNIPROT (\href{https://www.uniprot.org/proteomes/UP000000803})
}
\usage{
data(fasta)
}
\description{
File from UNIPROT containing protein sequences from Drosophila melanogaster. Exact fasta file used also in MaxQuant search
Contains sequence information of all proteins from Drosophila melanogaster in fasta format. Obtained from UNIPROT
}
\examples{

data(fasta)
head(fasta)

}
\keyword{fasta}
