% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.identical.R
\name{filter.identical}
\alias{filter.identical}
\title{filter.identical}
\usage{
\method{filter}{identical}(nonexclusive.data, exlusive.data)
}
\arguments{
\item{nonexclusive.data}{Data frame(s) (as list) obtained after limma differential expression analysis}

\item{exlusive.data}{Output from obtain_exclusive}
}
\value{
Data frame(s) (as list) for volcano plots
}
\description{
Remove identical proteins/sites from exclusive and non-exclusive data (only used when filter.protein.type = "fraction")
}
