% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDMR.R
\name{findDMR}
\alias{findDMR}
\title{Identify differentially methylated regions (DMR).}
\usage{
findDMR(
  expr = expr,
  sample = sample,
  samplename = samplename,
  group = group,
  ind = ind,
  method = "all"
)
}
\arguments{
\item{expr}{gene by cell normalized expression matrix.}

\item{sample}{a vector of sample names. Its length is the same as the number of columns in cell by gene normalized expression matrix.}

\item{samplename}{a vector the unique sample names}

\item{group}{a vector of the group indicator, 1 or 0, for differential analysis.}

\item{ind}{whether to construct design matrix.}

\item{method}{default is "all".  Other values are "limmacell_saver", "BSmooth_saver", "glmm_saver", "wilcoxon_saver", "t_saver", "scDD_saver", "MAST_saver", "limma_saver".}
}
\value{
test statistics and significance level
}
\description{
Identify differentially methylated regions (DMR).
}
\author{
Wenpin Hou <wp.hou3@gmail.com>
}
