% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAPlots.R
\name{GSEAPlots}
\alias{GSEAPlots}
\title{GSEAPlots}
\usage{
GSEAPlots(
  gseData,
  Fraction,
  enrich = c("gsea", "ora"),
  plotType = c("dotPlot", "cNetPlot", "heatPlot", "treePlot", "gseaPlot", "ridgePlot"),
  circular = FALSE,
  colorEdge = FALSE,
  nodeLabel = c("gene", "category", "all", "none"),
  cexLabelCategory = 1.2,
  cexLabelGene = 0.8,
  colorCcategory = "black",
  colorGene = "black",
  showCategory = 10,
  org = "dme",
  pvalCutOff = 0.05
)
}
\arguments{
\item{gseData}{Output from clusterProfiler (as list)}

\item{Fraction}{Can be either "Proteome" or "Enriched". Indicates the type of input data used}

\item{plotType}{Can be one of "dotPlot", "cNetPlot", "heatPlot", "treePlot", "gseaPlot", "ridgePlot"}

\item{circular}{Logical, default is FALSE. If TRUE, circular visualization will be used for cNetPlot}

\item{colorEdge}{Logical, default is FALSE. Parameter to be used if circular = "TRUE" and plotType ' "cNetPlot}

\item{nodeLabel}{Can be one of "gene", "category", "all", "none". Parameter to be used if circular = "TRUE" and plotType ' "cNetPlot}

\item{cexLabelCategory}{Parameter to be used if circular = "TRUE" and plotType ' "cNetPlot (Check enrichplot package for further info). Defaults i 1.2}

\item{cexLabelGene}{Parameter to be used if circular = "TRUE" and plotType ' "cNetPlot (Check enrichplot package for further info). Default is 0.8}

\item{colorCcategory}{Parameter to be used if circular = "TRUE" and plotType ' "cNetPlot (Check enrichplot package for further info). Default is "black"}

\item{colorGene}{Parameter to be used if circular = "TRUE" and plotType ' "cNetPlot (Check enrichplot package for further info). Default is "black"}

\item{showCategory}{Number of GO terms to be displayed in the plot. Default is 10}

\item{org}{Database of the organism. Drosophila melanogaster = "dme", Mus muscuslus ' "mmu", Homo sapiens = "hsa", Saccharomyces cerevisae = "sce". Default is "dme"}

\item{pvalCutOff}{P-value cut off for significant protein(s), site(s) and GO terms. Default is 0.05}
}
\value{
Generates plots from GO term analysis
}
\description{
Visualization of GSEA or ORA
}
