% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap}
\alias{heatmap}
\title{heatmap}
\usage{
heatmap(
  data_impute,
  lfq.data = lfq.data,
  filter.protein.type = filter.protein.type,
  Fraction,
  contrasts,
  distance.matrix = c("spearman", "pearson", "uncentered correlation",
    "absolute pearson", "sqrt", "weird"),
  exclusive.data = exclusive.data,
  clustering.method = c("ward.D", "ward.D2", "single", "complete", "average", "mcquitty",
    "median", "centroid"),
  sampleTable,
  title = NA,
  pvalCutOff = 0.05,
  sigmaCutOff = 0.05,
  lfcCutOff = 0,
  org = "dme"
)
}
\arguments{
\item{data_impute}{Output from plot_pca function}

\item{lfq.data}{Title of the plot}

\item{filter.protein.type}{Parameters (column should be present in the sampleTable file) to compare in the PCA (i.e. conditions or batch or timepoint)}

\item{Fraction}{Can be either "Proteome" or "Enriched". Indicates the type of input data used}

\item{contrasts}{Mentions the conditions to be compared. Ex. MUTANT_vs_WILDTYPE or (MUTANT-A_vs_WILDTYPE-A)_vs_(MUTANT-B_vs_WILDTYPE-B). This how the contrasts should be provided and the values should be the same as the one given in condition column of sampleTable. (MUTANT-A_vs_WILDTYPE-A)_vs_(MUTANT-B_vs_WILDTYPE-B): This type can be used for complex data when comparing the interaction between two conditions such genotype and time}

\item{distance.matrix}{One of "spearman", "pearson", "uncentered correlation", "absolute pearson", "sqrt", "weird"}

\item{exclusive.data}{Data frame(s) (as list) containing exclusive proteins or sites}

\item{clustering.method}{One of "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid"}

\item{sampleTable}{.xlsx file containing information about the samples. Three columns are mandatory (label, condition and replicate)}

\item{title}{Title of the plot}

\item{pvalCutOff}{P-value cut off for significant protein(s), site(s) and GO terms. Default is 0.05}

\item{sigmaCutOff}{PI-value cut off for significant protein(s) and site(s). Default is 0.05 (Refer Xiao et al, 2014 and Hostrup et al, 2022 for details on PI-value)}

\item{lfcCutOff}{Log-fold change cut off. Default is 0}

\item{org}{Database of the organism. Drosophila melanogaster = "dme", Mus muscuslus ' "mmu", Homo sapiens = "hsa", Saccharomyces cerevisae = "sce". Default is "dme"}
}
\value{
Generates heatmap and a file containing the order of protein(s) or site(s) in the heatmap
}
\description{
Generates heatmap from pheatmap
}
