% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_se_edit.R
\name{make_se_edit}
\alias{make_se_edit}
\title{make_se_edit}
\usage{
make_se_edit(proteins_unique, columns, expdesign)
}
\arguments{
\item{proteins_unique}{data containing the list of proteins with unique protein names (output from make_unique function)}

\item{columns}{LFQ or iBAQ}

\item{expdesign}{Sample table containing at least label, condition and replicate column}
}
\value{
An object of SummarizedExperiment class
}
\description{
Generate summarized experiment object (Obtained and edited from DEP package)
}
