% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestgenepredict.R
\name{nearestgenepredict}
\alias{nearestgenepredict}
\title{Predict DNA methylation using nearest gene's expression.}
\usage{
nearestgenepredict(trainexpr, testexpr, trainmeth, species = "human")
}
\arguments{
\item{trainexpr}{gene by sample expresion matrix for model training}

\item{testexpr}{gene by sample expresion matrix for model prediction}

\item{trainmeth}{CpG by sample DNA methylation matrix for model training}

\item{species}{character. "human" (default) or "mouse".}
}
\value{
predicted DNA methylation signals.
}
\description{
Predict DNA methylation using nearest gene's expression.
}
\author{
Wenpin Hou <wp.hou3@gmail.com>
}
