% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_exclusive.R
\name{obtain_exclusive}
\alias{obtain_exclusive}
\title{obtain_exclusive}
\usage{
obtain_exclusive(data)
}
\arguments{
\item{data}{output file (or output) from editData function}
}
\value{
List of exclsive proteins/sites from all comparisons (except difference of difference comparisons)
}
\description{
Get exclusive proteins/sites (only used when filter.protein.type = "fraction")
}
