% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Perform the model pretraining.}
\usage{
predict(expr, model, impute = F)
}
\arguments{
\item{expr}{gene by sample expresion matrix for model prediction.}

\item{model}{the model that has been pre-trained.}

\item{impute}{logical. Default=F. Impute the gene expression or not.}
}
\value{
a predicted DNA methylation matrix of beta values for CpG (rows) by sample (columns)
}
\description{
Perform the model pretraining.
}
\author{
Wenpin Hou <wp.hou3@gmail.com>
}
