% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prot.Data}
\alias{prot.Data}
\title{Input proteome data}
\format{
An object of class \code{"data.frame"}
}
\source{
\href{https://www.embopress.org/doi/full/10.15252/embr.202357023}{EMBO Reports}
}
\usage{
data(prot.Data)
}
\description{
This file is obtained from MaxQuant as proteinGroups.txt. Column information is not explained as these can be
obtained in the summary.pdf file from MaxQuant output. This corresponding input proteome data is obtained from a
recently published study comparing the whole cell extracts of proteomes between wild-type and chameau RNAi flies
(Further details can be found in the link provided below)
}
\examples{

data(prot.Data)
head(prot.Data)

}
\references{
Venkatasubramani AV, et al., (2023), EMBO Rep, 24(10):e57023
}
\keyword{input}
