% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainmodel.R
\name{trainmodel}
\alias{trainmodel}
\title{Perform the model pretraining.}
\usage{
trainmodel(
  expr,
  meth,
  log.transform = TRUE,
  filter.low.express.gene = TRUE,
  filter.low_var.gene = TRUE
)
}
\arguments{
\item{expr}{gene by sample expresion matrix for model training}

\item{meth}{CpG by sample DNA methylation matrix for model training}

\item{log.transform}{logical. default = TRUE. Whether to perform logit-transformation to the DNA methylation data}

\item{filter.low.express.gene}{logical. default = TRUE. Whether to filter out lowly expressed genes.}

\item{filter.low_var.gene}{logical. default = TRUE. Whether to filter out lowly variable genes.}
}
\value{
a list of pretrained model paramters
}
\description{
This function is used to perform the model pretraining.
}
\author{
Wenpin Hou <wp.hou3@gmail.com>
}
