% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainmodel_cvcluster.R
\name{trainmodel_cvcluster}
\alias{trainmodel_cvcluster}
\title{Perform the model pretraining.}
\usage{
trainmodel_cvcluster(expr, meth)
}
\arguments{
\item{expr}{gene by sample expresion matrix for model training}

\item{meth}{CpG by sample DNA methylation matrix for model training}
}
\value{
a list of pretrained model paramters
}
\description{
Perform the model pretraining.
}
\author{
Wenpin Hou <wp.hou3@gmail.com>
}
