% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainpredict.R
\name{trainpredict}
\alias{trainpredict}
\title{Perform the model training and prediction.}
\usage{
trainpredict(
  trainexpr,
  testexpr,
  trainmeth,
  clunumlist = c(1000, 2500, 5000),
  lambdalist = c(10^c(-1, 0, 1))
)
}
\arguments{
\item{trainexpr}{gene by sample expresion matrix for model training}

\item{testexpr}{gene by sample expresion matrix for model testing}

\item{trainmeth}{CpG by sample DNA methylation matrix for model training}

\item{clunumlist}{number of clusters for the genes, selected from c(1000,2500,5000)}

\item{lambdalist}{lambda value for Ridge regression, selected from  c(10^c(-1,0,1))}
}
\value{
a predicted DNA methylation matrix of beta values for CpG (rows) by sample (columns)
}
\description{
This function is used to perform the model fitting.
}
\examples{
data(demo_expr)
}
\author{
Wenpin Hou <wp.hou3@gmail.com>
}
