% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.timePlots.R
\name{bar.timePlots}
\alias{bar.timePlots}
\title{bar.timePlots}
\usage{
bar.timePlots(imputed.data, fav.proteins, timeSeries = FALSE, lfq.data)
}
\arguments{
\item{imputed.data}{Imputed data obtained after QC.filter}

\item{fav.proteins}{Vector with favourite proteins (or their corresponding sites) for which corresponding bar or timeseries plots are required}

\item{timeSeries}{Logical (Default is FALSE). If TRUE, line plots will be generated instead of barplots}

\item{lfq.data}{Output from editData function}
}
\value{
Generates bar plot (or timeseries) for the selected protein(s) or site(s)
}
\description{
Generates barplots or timeseries plots
}
