% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.identical.R
\name{filter.identical}
\alias{filter.identical}
\title{filter.identical}
\usage{
\method{filter}{identical}(nonexclusive.data, exlusive.data, Fraction, contrasts)
}
\arguments{
\item{nonexclusive.data}{Data frame(s) (as list) obtained after limma differential expression analysis}

\item{exlusive.data}{Output from obtain_exclusive}

\item{Fraction}{Can be either "Proteome" or "Enriched". Indicates the type of input data used}

\item{contrasts}{Mentions the conditions to be compared. Ex. MUTANT_vs_WILDTYPE or (MUTANT-A_vs_WILDTYPE-A)_vs_(MUTANT-B_vs_WILDTYPE-B). This how the contrasts should be provided and the values should be the same as the one given in condition column of sampleTable. (MUTANT-A_vs_WILDTYPE-A)_vs_(MUTANT-B_vs_WILDTYPE-B): This type can be used for complex data when comparing the interaction between two conditions such genotype and time}
}
\value{
Data frame(s) (as list) for volcano plots
}
\description{
Remove identical proteins/sites from exclusive and non-exclusive data (only used when filter.protein.type = "fraction")
}
