% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap}
\alias{heatmap}
\title{heatmap}
\usage{
heatmap(
  data_impute,
  lfq.data = lfq.data,
  distance.matrix = c("spearman", "pearson", "uncentered correlation",
    "absolute pearson", "sqrt", "weird"),
  exclusive.data = exclusive.data,
  clustering.method = c("ward.D", "ward.D2", "single", "complete", "average", "mcquitty",
    "median", "centroid"),
  title = NA
)
}
\arguments{
\item{data_impute}{Output from plot_pca function}

\item{lfq.data}{Title of the plot}

\item{distance.matrix}{One of "spearman", "pearson", "uncentered correlation", "absolute pearson", "sqrt", "weird"}

\item{exclusive.data}{Data frame(s) (as list) containing exclusive proteins or sites}

\item{clustering.method}{One of "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid"}

\item{title}{Title of the plot}
}
\value{
Generates heatmap and a file containing the order of protein(s) or site(s) in the heatmap
}
\description{
Generates heatmap from pheatmap
}
