% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.analysis.R
\name{motif.analysis}
\alias{motif.analysis}
\title{motif.analysis}
\usage{
motif.analysis(
  raw.enrichData,
  Ex.data,
  nonEx.data,
  fasta,
  aa = "K",
  seq.width = 15,
  min.seqs = 20,
  p.value.motif = 1e-05
)
}
\arguments{
\item{Ex.data}{Data frame(s) (as list) containing exclsive sites (generated only of filter.protein.type = "fraction")}

\item{nonEx.data}{Data frame(s) (as list) containing non-exclusive sites (Obtained after limma differential expression analysis)}

\item{fasta}{fasta file from uniprot (same fasta flies used for MaxQuant search)}

\item{aa}{Central amino acid. For example, "K" (for acetylome), "S", "T", "Y", "STY" (for phosphoproteome). Default is "K".}

\item{seq.width}{Width of the sequence for motif search. Default is 15}

\item{min.seqs}{This threshold refers to the minimum number of times you wish each of your extracted motifs to occur in the data set. Default is 20 and is usually appropriate, although this parameter may be adjusted to yield more specific or less specific motifs.}

\item{p.value.motif}{The p-value threshold for the binomial probability. This is used for the selection of significant residue/position pairs in the motif. Default is 1e-05 and is suggested to maintain a low false positive rate in standard protein motif analyses.}
}
\value{
Image of the most representated motif(s) and the corresponding information as a file (Check rmotifx package for more information)
}
\description{
Generates motifs using the peptide information of the enriched data (both exclusive and non-exclusive significant ones) using rmotifx
}
