% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.RAMP.R
\name{print.RAMP}
\alias{print.RAMP}
\title{Result summary of a fitted RAMP object.}
\usage{
\method{print}{RAMP}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{Fitted \code{'RAMP'} model object.}

\item{digits}{The number of significant digits for the coefficient estimates.}

\item{\dots}{Not used. Other arguments to predict.}
}
\value{
No value is returned.
}
\description{
Similar to the usual print methods, this function summarize results 
from a fitted \code{'RAMP'} object.
}
\seealso{
\code{\link{RAMP}}
}
