% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{volcanoPlot}
\usage{
volcanoPlot(proteinList, name.sigProteins = FALSE, resData)
}
\arguments{
\item{proteinList}{Vector with favourite proteins (or their corresponding sites) to be named in the volcano plot irrespective of its statistical significance}

\item{name.sigProteins}{Logical (Default is FALSE). If TRUE, names all significant protein(s) or site(s)}

\item{resData}{Data frame(s) (as list) from limma differential expression analysis}
}
\value{
Generates volcanoPlot
}
\description{
Generates heatmap from pheatmap
}
